Add physics body and contact test to Shark, Add extensions to GameScene
